Crawler agent factory reduce dependency from loader
